Add a wait to IA governor form config
It seems that the IA governor form uses a spam-detection technique of checking how long was spent between visiting the form and actual form submission. 5 seconds plus the additional time needed to fill out the form should be enough to not be detected as spam, so add a 5 second wait.